requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
kafka-python==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.4.2
requests==2.31.0
python-dotenv==1.0.0